upgrade to spring-boot 2.5.8 and telegrambots 6.0.1 - change data types of some telegram API entity fields - disable telegram by default (as it requires real bot token/username) - add empty default values for telegram properties - fix test assertions to JUnit5 - add re-routing messages to user in test_local profile (see *Spy)

+++ src/main/java/org/vors/pairbot/model/UserInfo.kt
@@ -11,7 +11,7 @@ import javax.persistence.*
 @org.hibernate.annotations.Cache(usage = CacheConcurrencyStrategy.READ_WRITE)
 @Table(name = "users")
 class UserInfo(
-        var userId: Int,
+        var userId: Long,
         var firstName: String,
 
         var lastName: String? = null,
